+++ AndroidStudioProject/app/src/main/java/org/gkisalatiga/plus/global/GlobalSchema.kt
@@ -161,6 +161,7 @@ class GlobalSchema : Application() {
         var fragmentServicesScrollState: ScrollState? = null
         var fragmentInfoScrollState: ScrollState? = null
         var screenAgendaScrollState: ScrollState? = null
+        var screenPersembahanScrollState: ScrollState? = null
 
         /* The poster dialog state in FragmentHome. */
         val fragmentHomePosterDialogState = mutableStateOf(false)